Implement eval_config
* rename init_settings method to init_config
* implement eval_config
* pass settings dictionary to commands instead of the config handler
  object

A note on terminology: in this context, we use the name config to
highlight the fact that the variable denotes a ConfigHandler object,
while settings is a combination of these config values and the passed
command line arguments (args).

The eval_config method was implemented to move this logic to a single
place, thus reducing the likelihood that sometime in the future an
error will be made here due to prior code repetition.