Add security section to config, with option hash
* move print_diff and copy_to_clipboard functions to utils namespace
* read hash algorithm from command line or config file

Note: I first thought about a string to class instance conversion in
order to pass the algorithm parameter to the checksum compute function,
but using a globals lookup was ugly, hence the str_to_hash_algorithm
function. I don't really like it, so perhaps it's better to change the
signature of the compute function to a string in future? Though this
switch case would still end up there, one way or the other...